Export compact.py fixes to CASE-Utilities-Python
This patch is ready to merge into the feature branch
CASE-Utilities-Python-next when a release is issued with this pull
request merged:
https://github.com/casework/CASE-Utilities-Python/pull/2

(This is another place where compact.py had been updated out of sync
with what was loaded into CASE-Utilities-Python.  The GNU-Time
repository's copy was in sync. with the ExifTool repository's copy.)

References:
* [AC-154] XML Schema datatype prefix needs to be explicit in all
  examples

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>